+++ app/src/main/java/hexlet/code/app/service/UserService.java
@@ -1,28 +1,43 @@
 package hexlet.code.app.service;
 
-import hexlet.code.app.dto.UserCreateDTO;
-import hexlet.code.app.dto.UserDTO;
-import hexlet.code.app.dto.UserUpdateDTO;
+import hexlet.code.app.dto.user.UserCreateDTO;
+import hexlet.code.app.dto.user.UserDTO;
+import hexlet.code.app.dto.user.UserUpdateDTO;
+import hexlet.code.app.exception.MethodNotAllowedException;
 import hexlet.code.app.exception.ResourceNotFoundException;
 import hexlet.code.app.mapper.UserMapper;
+import hexlet.code.app.repository.TaskRepository;
 import hexlet.code.app.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AllArgsConstructor;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
+
+import static hexlet.code.app.utils.Utils.urlNormalize;
 
 @Service
-public class UserService {
-    @Autowired
+@AllArgsConstructor
+public class UserService implements UserDetailsService {
     private UserRepository repository;
-    @Autowired
     private UserMapper mapper;
+    private TaskRepository taskRepository;
 
-    public List<UserDTO> getAll() {
-        var users = repository.findAll();
+    public List<UserDTO> getAll(Map<String, String> allParams) {
+        Map<String, Object> params = urlNormalize(allParams);
+        var users = repository.findAll(PageRequest.of((Integer) params.get("page"),
+                (Integer) params.get("perPage")));
         return users.stream()
                 .map(p -> mapper.map(p))
                 .toList();
+    }
+
+    public int count() {
+        return repository.countBy();
     }
 
     public UserDTO getById(long id) {
@@ -46,6 +61,15 @@ public class UserService {
     }
 
     public void delete(long id) {
+        if (taskRepository.existsTasksByAssignee_Id(id)) {
+            throw new MethodNotAllowedException("There are tasks associated with this assignee");
+        }
         repository.deleteById(id);
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        return repository.findByEmail(email)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
     }
 }
